Refactored Payments Domain Structure into slices
