+++ Sample/ECommerce/Payments/Payments.Api/Requests/Carts/DiscardPaymentRequest.cs
@@ -1,4 +1,4 @@
-using Payments.Payments.Enums;
+using Payments.Payments.DiscardingPayment;
 
 namespace Payments.Api.Requests.Carts
 {
