--- Sample/ECommerce/Payments/Payments.Tests/Extensions/AggregateExtensions.cs
@@ -1,14 +1,0 @@
-// using System.Linq;
-// using Core.Aggregates;
-// using Core.Events;
-//
-// namespace Payments.Tests.Extensions
-// {
-//     public static class AggregateExtensions
-//     {
-//         public static T PublishedEvent<T>(this IAggregate aggregate) where T : class, IEvent
-//         {
-//             return aggregate.DequeueUncommittedEvents().LastOrDefault() as T;
-//         }
-//     }
-// }
