--- Sample/ECommerce/Payments/Payments.Tests/Stubs/Ids/FakeIdGenerator.cs
@@ -1,12 +1,0 @@
-// using System;
-// using Core.Aggregates;
-// using Core.Ids;
-//
-// namespace Payments.Tests.Stubs.Ids
-// {
-//     public class FakeIdGenerator : IIdGenerator
-//     {
-//         public Guid? LastGeneratedId { get; private set; }
-//         public Guid New() => (LastGeneratedId = Guid.NewGuid()).Value;
-//     }
-// }
