--- Sample/ECommerce/Payments/Payments.Tests/Stubs/Reservations/FakeReservationNumberGenerator.cs
@@ -1,11 +1,0 @@
-// using System;
-// using Payments.Reservations;
-//
-// namespace Payments.Tests.Stubs.Reservations
-// {
-//     internal class FakeReservationNumberGenerator: IReservationNumberGenerator
-//     {
-//         public string LastGeneratedNumber { get; private set; }
-//         public string Next() => LastGeneratedNumber = Guid.NewGuid().ToString();
-//     }
-// }
