+++ Sample/ECommerce/Payments/Payments/Payments/RequestingPayment/PaymentRequested.cs
@@ -2,7 +2,7 @@ using System;
 using Ardalis.GuardClauses;
 using Core.Events;
 
-namespace Payments.Payments.Events
+namespace Payments.Payments.RequestingPayment
 {
     public class PaymentRequested: IEvent
     {
